requirements.txt
beautifulsoup4==4.9.3
boto3==1.15.16
botocore==1.18.16
certifi==2020.6.20
chardet==3.0.4
idna==2.10
jmespath==0.10.0
lyricsgenius==2.0.2
numpy==1.19.2
pandas==1.1.3
protobuf==3.13.0
PyMySQL==0.10.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
spotipy==2.16.0
SQLAlchemy==1.3.20
urllib3==1.25.10
